Fix int overflow on large torrents
svn path=/trunk/bitsharp/; revision=114048

+++ src/MonoTorrent/MonoTorrent.Client/ClientEngine.cs
@@ -291,7 +291,7 @@ namespace MonoTorrent.Client
         {
             get
             {
-                return Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.DownloadSpeed; });
+                return (int)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.DownloadSpeed; });
             }
         }
 
@@ -299,7 +299,7 @@ namespace MonoTorrent.Client
         {
             get
             {
-                return Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.UploadSpeed; });
+                return (int)Toolbox.Accumulate<TorrentManager>(torrents, delegate(TorrentManager m) { return m.Monitor.UploadSpeed; });
             }
         }
